+++ app/services/projects_for_export_service.rb
@@ -9,7 +9,16 @@ class ProjectsForExportService
     AcademiesApiPreFetcherService.new.call!(projects)
   end
 
+  def grant_management_and_finance_unit_projects(month:, year:)
+    projects = projects_by_advisory_board_date(month, year)
+    AcademiesApiPreFetcherService.new.call!(projects)
+  end
+
   private def projects_by_month_and_year(month, year)
     Conversion::Project.confirmed.filtered_by_significant_date(month, year)
+  end
+
+  private def projects_by_advisory_board_date(month, year)
+    Conversion::Project.confirmed.filtered_by_advisory_board_date(month, year)
   end
 end